Move logger into common
