--- aiohttp_json_api/log.py
@@ -1,7 +1,0 @@
-"""
-Logger instance
-===============
-"""
-import logging
-
-logger = logging.getLogger('aiohttp-json-api')
